requirements.txt
pandas
celery
bs4
sqlalchemy
sqlite3
redis
python-decouple
pytest
requests
eventlet